added a Basic Headtracking Space
the headtracking is alligned towards a target cross, Roll Pitch and Yaw get aligned with the height of the head. the desired steps are pitch and Yaw to a 16 and 32 degree , and Roll in 10 and 15 degree.

There is currently no implemented functions. Probably we use similar functions as in LoopAR. Probably we need to find a solution if the participant should allign itself or this is done with help of the experimenter.